requirements.txt
hdx-python-api==6.2.0

slugify~=0.0.1
